+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# [next]
+- Fix bug run when not visible [`AutomaticRandomMovement`]
+
 # [2.10.6]
 - Improvements in `AutomaticRandomMovement`
 - Improvements in `Follower`